Added structure to handle new pages
